moved calculatePosition to a separate file
